tests/wpt/css-tests/cssom-view-1_dev/html/window-screen-height.htm

<!DOCTYPE html>
<html><head>
    <title>CSSOM View - 4.2 - screen.height range tests</title>
    <link href="mailto:neils.christoffersen@gmail.com" rel="author" title="Neils Christoffersen">
	<link href="http://www.w3.org/TR/cssom-view/#the-screen-interface" rel="help">
    <meta content="dom" name="flags">
    <meta content="window.screen.height has sensible values" name="assert">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="myDiv"></div>
    <div id="log"></div>
    <script>
		var upperBound = 50000;

        test(
			function() {
				assert_not_equals(window.screen.height, 0, "window.screen.height is zero")
			}, "zero check"
		);

		test(
			function() {
				assert_true(window.screen.height > 0, "window.screen.height shouldn't be negative")
			}, "positive check"
		);

		test(
			function() {
				assert_true(window.screen.height < upperBound, "window.screen.height shouldn't be so large")
			}, "upper bound check"
		);
    </script>


</body></html>
